chore: seperate types into seperate module
